管理员 公告 投诉 个人信息

+++ pms-form/pms-form-api/src/main/java/com/example/form/controller/PictController.java
@@ -1,6 +1,7 @@
 package com.example.form.controller;
 
 import com.example.config.AuthThreadLocal;
+import com.example.exception.PMSException;
 import com.example.form.model.po.Pict;
 import com.example.form.service.PictService;
 import com.example.model.RestResponse;
@@ -13,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 
@@ -36,7 +38,7 @@ public class PictController {
      */
     @ApiOperation("文件上传api")
     @RequestMapping(value = "/upload", method = RequestMethod.POST)
-    public RestResponse<Pict> uploadPict(@RequestPart("file") MultipartFile file) throws IOException {
+    public RestResponse<Pict> uploadPict(@RequestPart("file") MultipartFile file, HttpServletRequest request) throws IOException {
         ThreadLocal<UserThreadLocalDto> authThreadLocal = AuthThreadLocal.getAuthThreadLocal();
         IsAuth.init(authThreadLocal).or("910").or("900").start();
 
@@ -47,8 +49,13 @@ public class PictController {
         String path = temp.getAbsolutePath();
         log.info("上传的文件路径为: {}", path);
 
-        //400表示投诉信息
-        RestResponse<Pict> pictRestResponse = pictService.uploadPict(file.getOriginalFilename(), path, "400");
+        String type = request.getHeader("type");
+        log.info("从headers传入的的值为： {}", type);
+
+        //type表示文件所属信息
+        if (type == null)
+            throw new PMSException("没有携带指定header");
+        RestResponse<Pict> pictRestResponse = pictService.uploadPict(file.getOriginalFilename(), path, type);
         if (temp.delete()) log.info("临时文件: {} 删除成功", file.getOriginalFilename());
         return pictRestResponse;
     }
